pyproject.toml
[tool.poetry]
name = "bot-test"
version = "0.1.0"
description = ""
authors = ["lin_vik <leogerte3@gmail.com>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.21.0"
sqlalchemy = "^2.0.41"
inflect = "^7.5.0"
pydantic-settings = "^2.10.1"
aiosqlite = "^0.21.0"
colorlog = "^6.9.0"
celery = "^5.5.3"
redis = "^6.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

